Save section jam to dabase

+++ app/src/main/java/pl/rubajticos/trafficjammer/domain/model/Models.kt
@@ -4,13 +4,13 @@ import java.time.LocalDateTime
 import java.util.*
 
 data class JamSection(
-    val id: Int? = null,
+    val id: Long? = null,
     val routeName: String,
     val trafficSignals: List<TrafficSignal>
 )
 
 data class TrafficSignal(
-    val id: Int? = null,
+    val id: Long? = null,
     val directionTo: String,
     val config: TrafficSignalConfig,
     val changes: List<TrafficSignalState>,
@@ -24,7 +24,7 @@ data class TrafficSignalConfig(
 )
 
 data class TrafficSignalState(
-    val id: Int? = null,
+    val id: Long? = null,
     val changeDate: LocalDateTime,
     val color: TrafficSignalColor,
     val executed: Boolean = false
